Rework structure.
