+++ src/components/StudentGroup/AddMembersPopup/styles.module.scss
@@ -1,0 +1,28 @@
+.wrapper {
+    overflow-y: scroll;
+    max-height: 80vh;
+}
+
+.addUserCardTitle {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    padding: 1em;
+
+    h2 {
+        margin: 0 0.4em;
+    }
+}
+
+.importFromList {
+    padding: 1em;
+
+    .textArea {
+        margin-top: 0.5em;
+        white-space: nowrap;
+        overflow-x: auto;
+        overflow-y: scroll;
+        width: 100%;
+        max-height: 500px;
+    }
+}
