+++ src/components/StudentGroup/AddMembersPopup/types.d.ts
@@ -1,0 +1,8 @@
+export interface AddMembersPopupProps {
+    studentGroup: StudentGroupModel;
+    onImported: (ids: string[], fromGroup?: StudentGroupModel) => void;
+}
+
+export type _AddMembersPopupPropsInternal = AddMembersPopupProps & {
+    popupRef: React.RefObject<PopupActions | null>;
+};
