+++ src/components/StudentGroup/styles.module.scss
@@ -100,32 +100,3 @@
     flex-direction: row;
     gap: 0.5em;
 }
-
-.wrapper {
-    overflow-y: scroll;
-    max-height: 80vh;
-}
-
-.addUserCardTitle {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 1em;
-
-    h2 {
-        margin: 0 0.4em;
-    }
-}
-
-.importFromList {
-    padding: 1em;
-
-    .textArea {
-        margin-top: 0.5em;
-        white-space: nowrap;
-        overflow-x: auto;
-        overflow-y: scroll;
-        width: 100%;
-        max-height: 500px;
-    }
-}
